[fix]修复首次安装加载本地服务资源的异常

+++ sdk/lib/src/files/file_util.dart
@@ -5,6 +5,9 @@ Future<bool> deleteDir(String path, {recursive: false}) async {
     throw Exception("This is a file path not a directory path");
   }
   var directory = Directory(path);
+  if (!directory.existsSync()) {
+    return false;
+  }
   try {
     await directory.delete(recursive: recursive);
     return true;
